spectral-analyzer-web/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Core dependencies
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
scipy>=1.11.0

# AI and API
openai==1.3.5
anthropic==0.7.1
httpx==0.25.2
aiohttp==3.9.1

# Data validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Caching and database
redis==5.0.1
aiosqlite==0.19.0

# File handling
chardet==5.2.0
openpyxl==3.1.2
python-docx==1.1.0

# Compression
lz4==4.3.2

# Monitoring
psutil==5.9.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS
fastapi-cors==0.0.6

# WebSockets
websockets==12.0

# Utilities
python-dotenv==1.0.0